External purchase button
Content can be configured with an external purchase link as a custom
field. This was a copypasta of the can be watched button with some
changes.

For styling, I've tried to piggyback off the existing purchase options
styling rather than introducing completely separate styling rules. This
looks more visually consistent than the can-be-watched button and
hopefully is a bit easier to style as a result.

I've tidied up the template uses of the hack buttons too so that they
behave consistently.

i18n: both buttons now get their labels i18n'd