Fix wrapping issue on collection views due to margin hack on first item
There was a -5px margin being added to the first child of a slider or
collection in order to counter-style the padding and align it with
titles -- except this looks goofy when it wraps because the posters
don't line up in a grid.

Instead, this applies the negative margin to the whole container.