Disable wishlist infinite scroll to fix jank when deleting items
Something about the infinite scroll hack was breaking dynamic sliders
when items were removed after the slider had been scrolled.

This is still not ideal because the slider scroll position seems to
reset, but I think that's a Relish bug caused by the way it updates the
list of partials.